делает модуляцию
